+++ particle_tracking_manager/the_manager.py
@@ -142,6 +142,11 @@ class ParticleTrackingManager:
         Name of input/output module type to use for writing Lagrangian model output. Default is "netcdf".
     use_cache : bool
         Set to True to use cache for saving interpolators, by default True.
+    interpolator_filename : Optional[Union[pathlib.Path,str]], optional
+        Filename to save interpolators to, by default None. The full path should be given, but no suffix.
+        Use this to either read from an existing file at a non-default location or to save to a
+        non-default location. If None and use_cache==True, the filename is set to a built-in name to an
+        `appdirs` cache directory.
 
     Notes
     -----
@@ -198,6 +203,9 @@ class ParticleTrackingManager:
         output_file: Optional[str] = config_ptm["output_file"]["default"],
         output_format: str = config_ptm["output_format"]["default"],
         use_cache: bool = config_ptm["use_cache"]["default"],
+        interpolator_filename: Optional[Union[pathlib.Path, str]] = config_ptm[
+            "interpolator_filename"
+        ]["default"],
         **kw,
     ) -> None:
         """Inputs necessary for any particle tracking."""
@@ -231,24 +239,16 @@ class ParticleTrackingManager:
 
         self.output_file_initial = None
 
-        # Set all attributes which will trigger some checks and changes in __setattr__
-        # these will also update "value" in the config dict
-        for key in sig.parameters.keys():
-            # no need to run through for init if value is None (already set to None)
-            if locals()[key] is not None:
-                self.__setattr__(key, locals()[key])
+        if output_file is None:
+            output_file = f"output-results_{datetime.datetime.now():%Y-%m-%dT%H%M:%SZ}"
 
-        self.kw = kw
-
-        if self.__dict__["output_file"] is None:
-            self.__dict__[
-                "output_file"
-            ] = f"output-results_{datetime.datetime.now():%Y-%m-%dT%H%M:%SZ}"
+        # want output_file to not include any suffix
+        output_file = output_file.rstrip(".nc").rstrip(".parq")
 
         ## set up log for this simulation
         # Create a file handler
-        assert self.__dict__["output_file"] is not None
-        logfile_name = self.__dict__["output_file"] + ".log"
+        assert output_file is not None
+        logfile_name = output_file + ".log"
         self.file_handler = logging.FileHandler(logfile_name)
         self.logfile_name = logfile_name
 
@@ -263,6 +263,56 @@ class ParticleTrackingManager:
 
         self.logger.info(f"filename: {logfile_name}")
         ##
+
+        if interpolator_filename is not None and not use_cache:
+            raise ValueError(
+                "If interpolator_filename is input, use_cache must be True."
+            )
+
+        # deal with caching/interpolators
+        # save interpolators to save time
+        if use_cache:
+            cache_dir = pathlib.Path(
+                appdirs.user_cache_dir(
+                    appname="particle-tracking-manager",
+                    appauthor="axiom-data-science",
+                )
+            )
+            cache_dir.mkdir(parents=True, exist_ok=True)
+            cache_dir = cache_dir
+            if interpolator_filename is None:
+                interpolator_filename = cache_dir / pathlib.Path(
+                    f"{ocean_model}_interpolator"
+                ).with_suffix(".pickle")
+            else:
+                interpolator_filename = pathlib.Path(interpolator_filename).with_suffix(
+                    ".pickle"
+                )
+            interpolator_filename = str(interpolator_filename)
+            self.save_interpolator = True
+            # if interpolator_filename already exists, load that
+            if pathlib.Path(interpolator_filename).exists():
+                self.logger.info(
+                    f"Loading the interpolator from {interpolator_filename}."
+                )
+            else:
+                self.logger.info(
+                    f"A new interpolator will be saved to {interpolator_filename}."
+                )
+        else:
+            self.save_interpolator = False
+            # this is already None
+            # self.interpolator_filename = None
+            self.logger.info("Interpolators will not be saved.")
+
+        # Set all attributes which will trigger some checks and changes in __setattr__
+        # these will also update "value" in the config dict
+        for key in sig.parameters.keys():
+            # no need to run through for init if value is None (already set to None)
+            if locals()[key] is not None:
+                self.__setattr__(key, locals()[key])
+
+        self.kw = kw
 
     def __setattr_model__(self, name: str, value) -> None:
         """Implement this in model class to add specific __setattr__ there too."""
@@ -383,7 +433,8 @@ class ParticleTrackingManager:
                 # by this point, output_file should already be a filename like what is
                 # available here, from OpenDrift (if run from there)
                 if self.output_file is not None:
-                    output_file = self.output_file.rstrip(".nc")
+                    output_file = self.output_file
+                    # output_file = self.output_file.rstrip(".nc")
                 else:
                     output_file = (
                         f"output-results_{datetime.datetime.now():%Y-%m-%dT%H%M:%SZ}"
@@ -476,29 +527,6 @@ class ParticleTrackingManager:
                     "setting `seed_seafloor` from True to False since now setting a non-None z value"
                 )
                 self.seed_seafloor = False
-
-            # save interpolators to save time
-            if name == "use_cache":
-                if value:
-                    cache_dir = pathlib.Path(
-                        appdirs.user_cache_dir(
-                            appname="particle-tracking-manager",
-                            appauthor="axiom-data-science",
-                        )
-                    )
-                    cache_dir.mkdir(parents=True, exist_ok=True)
-                    self.cache_dir = cache_dir
-                    self.interpolator_filename = cache_dir / pathlib.Path(
-                        f"{self.ocean_model}_interpolator"
-                    )
-                    self.save_interpolator = True
-                    self.logger.info(
-                        f"Interpolators will be saved to {self.interpolator_filename}."
-                    )
-                else:
-                    self.save_interpolator = False
-                    self.interpolator_filename = None
-                    self.logger.info("Interpolators will not be saved.")
 
             # if reader, lon, and lat set, check inputs
             if (